extract out components
