+++ main.js
@@ -3,8 +3,8 @@ import VueRouter from 'vue-router'
 
 Vue.use(VueRouter)
 
-import App from './App.vue'
-import Canon from './Canon.vue'
+import App from './components/App.vue'
+import Canon from './components/Canon.vue'
 
 const router = new VueRouter({
   mode: 'history',
